+++ client/Packages/com.beamable/CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Editor tooling for `SerializableDictionaryStringToSomething<T>` has context menu.
 - Expanded/Collapsed state represented by icons in `Theme Manager`
+- New customer registrations will send an analytic event to Beamable
 
 ### Fixed
 - The default uncaught Promise handler no longer throws `IndexOutOfBounds` errors in high failure cases.